Chargement de la taille du véhicule en cours de réalisation

+++ cpp/IGC_Tron/IGC_Tron/CartesianVector.h
@@ -15,7 +15,7 @@ namespace Utils
 
 struct CartesianVector
 {
-	typedef double CoordType;
+	typedef float CoordType;
 
 	CoordType x, y, z;
